Implement order placement, payment and shipping (with tracking). General refactor

+++ src/main/java/xyz/jiniux/aap/controllers/requests/FillStockRequest.java
@@ -6,5 +6,5 @@ import java.io.Serializable;
 
 public record FillStockRequest(
     @Min(1)
-    int quantity
+    long quantity
 ) implements Serializable { }
