+++ src/main/java/xyz/jiniux/aap/controllers/requests/SyncShoppingCartRequestItem.java
@@ -20,5 +20,5 @@ public record SyncShoppingCartRequestItem(
     String stockQuality,
 
     @Min(1)
-    int quantity
+    long quantity
 ) {}
